--- backend/app/routes/data_imports/__init__.py
@@ -1,5 +1,0 @@
-"""Data import route package"""
-
-from app.routes.data_imports.router import router
-
-__all__ = ["router"]
