+++ backend/app/routes/transactions/router.py
@@ -1,7 +1,8 @@
 """Transaction API routes"""
 import uuid
-from datetime import date
+from datetime import date, datetime
 from typing import Annotated
+from zoneinfo import ZoneInfo
 
 from fastapi import APIRouter, Depends, Query, status
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -9,6 +10,12 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.database import get_db
 from app.dependencies import get_current_user
 from app.models.user import User
+from app.schemas.firefly_import import (
+    FireflyBudgetImportRequest,
+    FireflyBudgetImportResponse,
+    FireflyTransactionImportRequest,
+    FireflyTransactionImportResponse,
+)
 from app.schemas.transaction import (
     CreateTransactionRequest,
     TransactionImportRequest,
@@ -20,7 +27,11 @@ from app.schemas.transaction import (
 from app.services.transactions.creation import create_transaction_and_get_response
 from app.services.transactions.deletion import delete_transaction_for_user
 from app.services.transactions.detail import get_transaction_response_for_user
-from app.services.transactions.imports import import_transactions
+from app.services.transactions.imports import (
+    import_firefly_budgets,
+    import_firefly_transactions,
+    import_transactions,
+)
 from app.services.transactions.listing import list_transaction_responses
 from app.services.transactions.overview import get_transactions_overview as get_transactions_overview_response
 from app.services.transactions.update import update_transaction_and_get_response
@@ -174,6 +185,61 @@ async def import_transaction_batch(
         Import summary containing created and skipped transaction counts
     """
     return await import_transactions(db, user, data)
+
+
+@router.post(
+    "/import/firefly",
+    response_model=FireflyTransactionImportResponse,
+    status_code=status.HTTP_201_CREATED,
+)
+async def import_firefly_transaction_rows(
+    data: FireflyTransactionImportRequest,
+    user: Annotated[User, Depends(get_current_user)],
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
+    """Import journal rows from a Firefly III data export
+
+    Rows between an imported account and an expense or revenue counterparty
+    become single transactions, while rows between two imported accounts
+    become transfer pairs. Rows that cannot convert are skipped and reported
+
+    Args:
+        data: Prepared Firefly III import payload from the frontend compiler
+        user: Authenticated user running the import
+        db: Active database session
+
+    Returns:
+        Import summary with converted, skipped, and created record counts
+    """
+    return await import_firefly_transactions(db, user, data)
+
+
+@router.post(
+    "/import/firefly/budgets",
+    response_model=FireflyBudgetImportResponse,
+    status_code=status.HTTP_201_CREATED,
+)
+async def import_firefly_budget_rows(
+    data: FireflyBudgetImportRequest,
+    user: Annotated[User, Depends(get_current_user)],
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
+    """Import budgets from a Firefly III data export
+
+    Each budget becomes a monthly base budget backdated to its period start,
+    with one period instance per month through today carrying the limit
+    amount that was in force for that month
+
+    Args:
+        data: Budgets derived from the export by the frontend
+        user: Authenticated user running the import
+        db: Active database session
+
+    Returns:
+        Summary of the created budgets and their materialized periods
+    """
+    today = datetime.now(ZoneInfo(user.tz)).date()
+    return await import_firefly_budgets(db, user, data, today)
 
 
 @router.post("", response_model=TransactionResponse, status_code=status.HTTP_201_CREATED)
