+++ frontend/src/pages/imports/firefly/hooks/useFireflyImportWorkflow.ts
@@ -6,7 +6,7 @@ import {
   useImportFireflyBudgets,
   useImportFireflyTransactions,
   type FireflyTransactionImportResponse,
-} from '@/api/dataImports'
+} from '@/api/fireflyImports'
 import { useInstitutions } from '@/api/institutions'
 import { waitForMilliseconds } from '@/utils/timing'
 import { CREATE_ACCOUNT_VALUE, CREATE_CATEGORY_VALUE } from '../../constants'
